Move existing codegen into v0 api (preparing to add new code)
