+++ honc-code-gen/src/index.ts
@@ -1,43 +1,31 @@
 import { instrument } from "@fiberplane/hono-otel";
-import { zValidator } from "@hono/zod-validator";
 import { drizzle } from "drizzle-orm/d1";
 import { Hono } from "hono";
 import { logger as honoLogger } from "hono/logger";
-import { z } from "zod";
 import * as schema from "./db/schema";
-import {
-  FILES_TO_MODIFY,
-  // NOTE - Had to swap out Anthropic for OpenAI because of rate limiting concerns
-  // buildWithAnthropic,
-  buildWithAnthropicMock,
-} from "./lib/ai/anthropic";
-import { buildWithOpenAi } from "./lib/ai/openai";
-import { isScaffoldAppToolParameters } from "./lib/ai/tools";
-import {
-  readProjectFiles,
-  writeIndexFile,
-  writeSchemaFile,
-  writeSeedFile,
-} from "./lib/project-files";
-import type { Bindings, Variables } from "./lib/types";
+
+import type { HatchApp } from "./types";
 import { createLogger } from "./logger";
-import { SchemaContextSchema } from "./types";
+import v0Api from "./v0";
 
-const app = new Hono<{ Bindings: Bindings; Variables: Variables }>();
+const app = new Hono<HatchApp>();
 
+// Add a scoped logger on the context, since our logger depends on the env variables
 app.use(async (c, next) => {
   const logger = createLogger(c.env.HONC_LOG_LEVEL);
   c.set("appLogger", logger);
   await next();
 });
 
-// NOTE - This middleware adds `db` on the context so we don't have to initiate it every time
+// Add `db` on the context so we don't have to initiate it every time
 app.use(async (c, next) => {
   const db = drizzle(c.env.DB, { schema });
   c.set("db", db);
   await next();
 });
 
+// Add a global error handler
+// TODO - Use `app.onError`
 app.use(async (c, next) => {
   try {
     await next();
@@ -61,9 +49,9 @@ app.use(async (c, next) => {
   }
 });
 
-// Set up the builtin hono logger, but use debug logs from our logger
-// This means users of the Honc++ cli will not see the request logs in their terminal,
-// but if you want to see them locally, set `HONC_LOG_LEVEL=debug`
+// Set up the built-in hono logger, but use debug logs from our logger.
+// If you want to see debug logs locally, set `HONC_LOG_LEVEL=debug`,
+// when running this worker.
 app.use((c, next) =>
   honoLogger((message: string, ...rest: string[]) => {
     const logger = c.get("appLogger");
@@ -73,121 +61,6 @@ app.use((c, next) =>
 
 app.get("/", (c) => c.text("Hello, world!"));
 
-app.get("/v0/files/:sessionId", async (c) => {
-  const { sessionId } = c.req.param();
-  const files = await readProjectFiles(c.env.R2, sessionId);
-  return c.json({ files });
-});
-
-/**
- * Creates files for a given prompt, and save those files in R2, using the sessionId as the key.
- * Returns the list of files that were created.
- */
-app.post(
-  "/v0/build/:sessionId",
-  zValidator(
-    "json",
-    z.object({
-      prompt: z.string(),
-      indexFile: z.string().nullish(),
-      schemaFile: z.string().nullish(),
-      schemaContext: SchemaContextSchema.nullish(),
-      seedFile: z.string().nullish(),
-    }),
-  ),
-  async (c) => {
-    const logger = c.get("appLogger");
-
-    // HACK - Set this true to return a mock response, not eat up API calls
-    const USE_MOCK = false;
-    if (USE_MOCK) {
-      logger.warn("Using mock response for code generation");
-    }
-
-    const db = c.get("db");
-    const sessionId = c.req.param("sessionId");
-    const body = c.req.valid("json");
-    const { prompt, indexFile, schemaFile, schemaContext, seedFile } = body;
-
-    logger.debug("Project files:", {
-      indexFile: indexFile ? "present" : "missing",
-      schemaFile: schemaFile ? "present" : "missing",
-      seedFile: seedFile ? "present" : "missing",
-    });
-    logger.debug(
-      "Schema context:",
-      schemaContext
-        ? {
-            type: schemaContext?.type,
-            drizzleImport: schemaContext?.drizzleImport,
-            vendor: schemaContext?.vendor,
-          }
-        : "missing",
-    );
-
-    const builder = USE_MOCK ? buildWithAnthropicMock : buildWithOpenAi;
-
-    const result = await builder(
-      {
-        // apiKey: c.env.ANTHROPIC_API_KEY,
-        apiKey: c.env.OPENAI_API_KEY,
-        indexFile: indexFile || "<index file not provided />",
-        schemaFile: schemaFile || "<schema file not provided />",
-        schemaContext: schemaContext || null,
-        seedFile: seedFile || "<seed file not provided />",
-        userPrompt: prompt,
-      },
-      logger,
-    );
-
-    if (isScaffoldAppToolParameters(result)) {
-      // Don't await this, so we can return a response sooner and don't block the user's response
-      c.executionCtx.waitUntil(
-        Promise.all([
-          writeIndexFile(c.env.R2, sessionId, result.indexFile),
-          writeSchemaFile(c.env.R2, sessionId, result.schemaFile),
-          writeSeedFile(c.env.R2, sessionId, result.seedFile),
-        ]),
-      );
-    } else {
-      logger.error(
-        "Failed to parse scaffold app tool parameter according to schema",
-      );
-      await db
-        .insert(schema.sessions)
-        .values({
-          id: sessionId,
-          prompt,
-          data: {
-            result,
-            // NOTE - This is for future reference, if we change how and which files we modify
-            filesModified: FILES_TO_MODIFY,
-            schemaContext,
-          },
-          type: "error",
-        })
-        .returning();
-      return c.json(
-        { error: "Unparseable tool call response from Anthropic" },
-        500,
-      );
-    }
-
-    const [session] = await db
-      .insert(schema.sessions)
-      .values({
-        id: sessionId,
-        prompt,
-        data: {
-          // NOTE - This is for future reference, if we change how and which files we modify
-          filesModified: FILES_TO_MODIFY,
-          schemaContext,
-        },
-        type: "success",
-      })
-      .returning();
-    return c.json({ result, session });
-  },
-);
+app.route("/v0", v0Api);
 
 export default instrument(app);
