+++ honc-code-gen/src/v0/lib/ai/agent-example.ts
