+++ honc-code-gen/src/v0/lib/ai/openai.ts
