+++ honc-code-gen/src/v0/lib/types.ts
@@ -1,6 +1,5 @@
-// Only using for global types
-
 import type { DrizzleD1Database } from "drizzle-orm/d1";
+import { z } from "zod";
 import type * as schema from "../db/schema";
 import type { createLogger } from "../logger";
 
@@ -18,3 +17,16 @@ export type Variables = {
   db: DrizzleD1Database<typeof schema>;
   appLogger: ReturnType<typeof createLogger>;
 };
+
+export type HatchApp = {
+  Bindings: Bindings;
+  Variables: Variables;
+};
+
+export const SchemaContextSchema = z.object({
+  type: z.string(),
+  drizzleImport: z.string(),
+  vendor: z.string(),
+});
+
+export type SchemaContext = z.infer<typeof SchemaContextSchema>;
